Agregada lógica de restablecimiento de contraseña.
Implementar lógica de restablecimiento de contraseña en backend y frontend. Se crearon controladores y rutas para enviar correos electrónicos y permitir a los usuarios restablecer sus contraseñas olvidadas. Diseño, validaciones y funcionalidad de visibilidad de contraseña también fueron agregados.

+++ server/database/migrations/2023_11_02_232759_create_reports_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->unsignedBigInteger('user_id');
             $table->foreign('user_id')->references('id')->on('users');
             $table->string('month');
-            $table->float('income');
+            $table->decimal('income', 14, 2);
             $table->timestamps();
         });
     }
